requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-ratelimit==4.1.0
django-taggit==5.0.1
django-taggit-serializer==0.1.7
django_csp==3.8
djangorestframework==3.14.0
pillow==10.2.0
pymemcache==4.0.0
python-memcached==1.62
pytz==2024.1
ratelimit==2.2.1
six==1.16.0
sqlparse==0.4.4
